need to check BookRepositoryIT "pages Not-Null Constraint"

+++ catalog/src/main/java/mo/bfh/catalog/repository/BookRepository.java
@@ -3,6 +3,6 @@ package mo.bfh.catalog.repository;
 import mo.bfh.catalog.entity.Book;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BookRepository extends JpaRepository<Book, Integer> {
+public interface BookRepository extends JpaRepository<Book, Long> {
 
 }
